+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/changelog.py
@@ -14,8 +14,8 @@ from ...constants import CHANGELOG_TYPE_NONE, get_root
 from ...git import get_commits_since
 from ...github import from_contributor, get_changelog_types, get_pr, parse_pr_numbers
 from ...release import get_release_tag_string
-from ...utils import get_valid_checks, get_version_string, validate_check_arg
-from ..console import CONTEXT_SETTINGS, abort, echo_failure, echo_info
+from ...utils import get_valid_checks, get_version_string
+from ..console import CONTEXT_SETTINGS, abort, echo_failure, echo_info, validate_check_arg
 
 ChangelogEntry = namedtuple('ChangelogEntry', 'number, title, url, author, author_url, from_contributor')
